1.移除了PetContract，改用数据库存储 2.将金额和余额改为int类型 3.将sell_username改为owner 4.UserContract不再传入username，也不处理任何登录逻辑，改为传入address，登录逻辑在UserController和UerClient内处理 5.添加了User和Pet的数据库存储实现

+++ src/main/java/com/fisco/app/controller/UserController.java
@@ -63,7 +63,7 @@ public class UserController {
      * @return 如果有这个用户返回 ResponseData.success(“success") , 否则返回 ResponseData.error("can't find this user!")
      */
     @PostMapping("/add_balance")
-    public ResponseData add_balance(@RequestBody String username, @RequestBody double amount) {
+    public ResponseData add_balance(@RequestBody String username, @RequestBody int amount) {
         //调用UserClient内的函数
         boolean whether_add_success = userClient.add_balance(username, amount);
         if(whether_add_success)
